feat: Split renderer.ts (1,959 LOC) into three focused sub-modules unde…
- src/webview/render/html-builders.ts
- src/webview/render/streaming.ts
- src/webview/render/batches.ts
- src/webview/renderer.ts
- src/webview/__tests__/accessibility.test.ts

GSD-Task: S02/T03
